Makify

+++ toyllvm/test.c
@@ -1,3 +1,3 @@
-int add2(int base) {
+long add2(long base) {
 	return base + 2;
 }
